requirements.txt
torch 
torchvision
packaging
sentencepiece
einops
numpy
opencv_python
Pillow
pycocotools==2.0.6
Requests
tqdm
uvicorn
scipy
bitsandbytes
timm
blobfile
mypy
pytest
requests
tensorboardX
opencv-python
torchmetrics
deepspeed
pycocoevalcap
torchscale
hydra-core
accelerate
transformers
pyarrow
ftfy
tensorboard
datasets
diffusers
huggingface_hub
scikit-image
python-box
spacy
imageio
imageio[pyav]
imageio[ffmpeg]